test(esl-event-utils): update user cases tests for `decorate` utility
